예제 수정

+++ src/main/java/kopo/poly/service/impl/StudentService.java
@@ -46,7 +46,17 @@ public class StudentService implements IStudentService {
 
         log.info(this.getClass().getName() + ".updateStudent Start!");
 
-        studentMapper.updateStudent(pDTO); // 학생 수정 SQL 실행하기
+        // Student 테이블에 등록된 학생 아이디가 존재하는지 체크하기 위해 DB 조회하기
+        Optional<StudentDTO> res = Optional.ofNullable(studentMapper.getStudent(pDTO));
+
+        if (res.isPresent()) { // DB 조회 결과로 회원아이디가 존재한다면...
+            studentMapper.updateStudent(pDTO); // 학생 수정 SQL 실행하기
+            log.info(pDTO.getUserId() + "님이 수정되었습니다.");
+
+        } else {
+            log.info("회원이 존재하지 않아 수정되지 못했습니다.");
+
+        }
 
         // 학생 테이블 전체 조회하기
         List<StudentDTO> rList = Optional.ofNullable(studentMapper.getStudentList()).orElseGet(ArrayList::new);
@@ -61,7 +71,18 @@ public class StudentService implements IStudentService {
 
         log.info(this.getClass().getName() + ".deleteStudent Start!");
 
-        studentMapper.deleteStudent(pDTO); // 학생 삭제 SQL 실행하기
+        // Student 테이블에 등록된 학생 아이디가 존재하는지 체크하기 위해 DB 조회하기
+        Optional<StudentDTO> res = Optional.ofNullable(studentMapper.getStudent(pDTO));
+
+        if (res.isPresent()) { // DB 조회 결과로 회원아이디가 존재한다면...
+            studentMapper.deleteStudent(pDTO); // 학생 삭제 SQL 실행하기
+            log.info(pDTO.getUserId() + "님이 삭제되었습니다.");
+
+        } else {
+            log.info("회원이 존재하지 않아 삭제되지 못했습니다.");
+
+        }
+
 
         // 학생 테이블 전체 조회하기
         List<StudentDTO> rList = Optional.ofNullable(studentMapper.getStudentList()).orElseGet(ArrayList::new);
